refactor(frontend): reimport necessary components in their index.ts
