+++ src/App.tsx
@@ -4,7 +4,7 @@ import { RelayEnvironmentProvider } from 'react-relay/hooks';
 import { AppRouter } from 'src/AppRouter';
 import { RtlSupportProvider } from 'src/core/rtl/RtlSupportProvider';
 import { GlobalStyles } from 'src/core/styles/GlobalStyles';
-import { rtlTheme } from 'src/core/theme/themes';
+import { rtlTheme } from 'src/core/theme';
 import { environment } from 'src/relay';
 
 const App: React.FC = () => {
