+++ src/AppRouter.tsx
@@ -1,11 +1,11 @@
 import React, { Suspense } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import { AuthGuard } from 'src/core/auth/AuthGuard';
-import FullPageError from 'src/pages/full-page-error/FullPageError';
-import { FullPageSpinner } from 'src/pages/loading/FullPageSpinner';
+import { AuthGuard } from 'src/core/auth';
 import { ErrorBoundary } from 'src/shared/error-boundary';
+import { FullPageError } from 'src/shared/full-page-error';
+import { FullPageSpinner } from 'src/shared/loading';
 import { FCProps } from 'src/shared/types/FCProps';
-import { MainContainer } from './containers/main/MainContainer';
+import { MainContainer } from './containers/main';
 
 interface OwnProps {}
