+++ src/containers/main/index.ts
@@ -1,0 +1,1 @@
+export { MainContainer } from './MainContainer';
