+++ src/core/auth/index.ts
@@ -1,0 +1,1 @@
+export { AuthGuard } from './AuthGuard';
