+++ src/core/locales/index.ts
@@ -1,0 +1,1 @@
+export { getLanguages } from './languages';
