+++ src/core/locales/types.d.ts
@@ -1,0 +1,1 @@
+export { Language, LanguageCodes } from './languages';
