+++ src/core/rtl/index.ts
@@ -1,0 +1,1 @@
+export { RtlSupportProvider } from './RtlSupportProvider';
