+++ src/core/styles/index.ts
@@ -1,0 +1,1 @@
+export { GlobalStyles } from './GlobalStyles';
