+++ src/core/theme/index.ts
@@ -1,0 +1,1 @@
+export { darkLtrTheme, darkRtlTheme, ltrTheme, rtlTheme } from './themes';
