+++ src/shared/auth/index.ts
@@ -1,0 +1,1 @@
+export { AuthPage } from './AuthPage';
