+++ src/shared/counter/index.ts
@@ -1,0 +1,1 @@
+export { Counter } from './Counter';
