+++ src/shared/full-page-error/FullPageError.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 // TODO: create error page here
-export default function FullPageError() {
+export function FullPageError() {
   return <div>ERROR</div>;
 }
