+++ src/shared/full-page-error/index.ts
@@ -1,0 +1,1 @@
+export { FullPageError } from './FullPageError';
