+++ src/shared/loading/index.ts
@@ -1,0 +1,1 @@
+export { FullPageSpinner } from './FullPageSpinner';
